Change Oilfield to be with the new convention
